+++ rfs-pipeline/src/main/java/org/opensearch/migrations/bulkload/pipeline/ir/ProgressCursor.java
@@ -17,7 +17,7 @@ import java.util.Objects;
  */
 public record ProgressCursor(
     ShardId shardId,
-    int lastDocProcessed,
+    long lastDocProcessed,
     long docsInBatch,
     long bytesInBatch
 ) {
